2025.01.16.2
Refactor initialization and add SVG generation

Simplified the initialization of the `rotatedPoints` list by directly assigning the result of the `RotatePoints` method. Removed the `WriteLine` statement that printed the `rotatedPoints`. Added a new call to `SVGConverter.GenerateSVG` to generate an SVG file with the specified parameters and the `rotatedPoints` list.